requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.2
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
zipp==3.11.0
